Resize prediction screen. Move all the Twitch (Object) = in Entities folder
