+++ TwitchChannelPointsMiner/classes/entities/Bet.py
