+++ TwitchChannelPointsMiner/classes/entities/Message.py
