+++ TwitchChannelPointsMiner/classes/entities/Raid.py
