+++ TwitchChannelPointsMiner/classes/entities/Stream.py
